# sdk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChainSettle API",
    "description": "Backend API for ChainSettle — milestone-based supply chain escrow on Stellar Soroban. Regenerate with `npm run generate:sdk`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {},
          "timestamp": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["users"],
        "operationId": "UsersController_getProfile",
        "summary": "Get the authenticated user profile",
        "security": [{ "bearer": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Envelope" }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/login-history": {
      "get": {
        "tags": ["users"],
        "operationId": "UsersController_getLoginHistory",
        "summary": "Get recent sign-in history with IP and user agent",
        "security": [{ "bearer": [] }],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Login history records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "timestamp": { "type": "string", "format": "date-time" },
                          "ipAddress": { "type": "string", "nullable": true },
                          "userAgent": { "type": "string" },
                          "success": { "type": "boolean" },
                          "isCurrentSession": { "type": "boolean" }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "nextCursor": { "type": "string", "nullable": true },
                        "limit": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/avatar": {
      "put": {
        "tags": ["users"],
        "operationId": "UsersController_uploadAvatar",
        "summary": "Upload profile picture via IPFS",
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "avatarCid": { "type": "string" },
                    "avatarUrl": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "No file uploaded or unsupported MIME type" },
          "413": { "description": "File size exceeds 2 MB limit" }
        }
      },
      "delete": {
        "tags": ["users"],
        "operationId": "UsersController_deleteAvatar",
        "summary": "Delete user profile avatar",
        "security": [{ "bearer": [] }],
        "responses": {
          "200": {
            "description": "Avatar deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{stellarAddress}": {
      "get": {
        "tags": ["users"],
        "operationId": "UsersController_getPublicProfile",
        "summary": "Get a public user profile by Stellar address",
        "security": [{ "bearer": [] }],
        "parameters": [
          {
            "name": "stellarAddress",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Public user profile returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Envelope" }
              }
            }
          },
          "404": { "description": "User not found" }
        }
      }
    }
  }
}